+++ example/app/primeng/primeng-example.component.ts
@@ -7,7 +7,12 @@ import { PRIMENG_EXAMPLE_MODEL } from "./primeng-example.model";
 
     moduleId: module.id,
     selector: "dynamic-form-primeng-example",
-    styleUrls: ["../../../node_modules/primeng/resources/themes/omega/theme.css", "../../../node_modules/primeng/resources/primeng.min.css"],
+    styleUrls: [
+        "../../../node_modules/primeng/resources/themes/omega/theme.css",
+        "../../../node_modules/primeng/resources/primeng.min.css",
+        "../../../node_modules/quill/dist/quill.core.css",
+        "../../../node_modules/quill/dist/quill.snow.css",
+    ],
     templateUrl: "./primeng-example.component.html",
     encapsulation: ViewEncapsulation.None
 })